requirements.txt
discord.py
postgres
pychalk
python-dotenv